+++ tests/integ/test_dataframe.py
@@ -152,7 +152,6 @@ def test_dataframe_get_attr(session):
     assert "object has no attribute" in str(exc_info)
 
 
-# @pytest.mark.localtest
 @pytest.mark.skipif(IS_IN_STORED_PROC_LOCALFS, reason="need resources")
 def test_read_stage_file_show(session, resources_path, local_testing_mode):
     tmp_stage_name = Utils.random_stage_name()
@@ -160,8 +159,7 @@ def test_read_stage_file_show(session, resources_path, local_testing_mode):
     test_file_on_stage = f"@{tmp_stage_name}/testCSV.csv"
 
     try:
-        if not local_testing_mode:
-            Utils.create_stage(session, tmp_stage_name, is_temporary=True)
+        Utils.create_stage(session, tmp_stage_name, is_temporary=True)
         Utils.upload_to_stage(
             session, "@" + tmp_stage_name, test_files.test_file_csv, compress=False
         )
@@ -1578,7 +1576,7 @@ def test_create_dataframe_with_semi_structured_data_types(session):
 
 @pytest.mark.skipif(
     "config.getoption('local_testing_mode', default=False)",
-    reason="SNOW-1368516 create dataframe from pandas dataframe with datetime columns has wrong schema",
+    reason="SNOW-1439717 create dataframe from pandas dataframe containing timestamp with tzinfo is not supported.",
 )
 @pytest.mark.skipif(not is_pandas_available, reason="pandas is required")
 def test_create_dataframe_with_pandas_df(session):